backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
crewai
crewai-tools
spacy
litellm
groq